removed separate trigger/solid links. some q3 fixes. q2 will autosave on map changes, like q2 normally does.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@3839 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/cl_demo.c
@@ -167,7 +167,7 @@ int demo_preparsedemo(unsigned char *buffer, int bytes)
 		switch(buffer[1]&dem_mask)
 		{
 		case dem_cmd:
-			ofs = -(sizeof(q1usercmd_t));
+			ofs = -(int)(sizeof(q1usercmd_t));
 			ofs = 0;
 			break;
 		case dem_set:
